+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - `DATATYPE99_MAJOR`, `DATATYPE99_MINOR`, `DATATYPE99_PATCH`
 
+### Changed
+
+ - `Unit99` -> `UnitT99`, `unit99` -> `unit_v99` (less likely to conflict with user-defined identifiers)
+
 ## [0.4.0] - 2021-03-27
 
 ### Added